% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosHighlight.R
\name{circoshighlight}
\alias{circoshighlight}
\title{circosHighlight}
\usage{
circoshighlight(df, sectors, color, colorcat, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
highlight chart
}
